+++ bootstrap/generate.py
@@ -104,7 +104,7 @@ def create_module(
 
 import functools as functools__
 import numpy as numpy__
-from omc4py.primitive_types import (
+from omc4py.types import (
     TypeName,
     VariableName,
 )
