--- omc4py/primitive_types.py
@@ -1,222 +1,0 @@
-
-__all__ = (
-    "Boolean",
-    "Component",
-    "Integer",
-    "Real",
-    "String",
-    "TypeName",
-    "VariableName",
-)
-
-import numpy  # type: ignore
-import typing
-
-
-Real = numpy.double
-Integer = numpy.intc
-Boolean = numpy.bool
-String = numpy.str
-
-
-class VariableName(
-):
-    __slots__ = "__str"
-
-    __str: str
-
-    def __new__(
-        cls,
-        obj,
-    ):
-        if isinstance(obj, VariableName):
-            return obj
-
-        obj_str = str(obj)
-        if not parser.valid_identifier(obj_str):
-            raise ValueError(
-                f"Invalid modelica identifier, got {obj_str!r}"
-            )
-
-        return _VariableName_from_valid_identifier_no_check(
-            cls,
-            obj_str
-        )
-
-    def __eq__(
-        self, other,
-    ):
-        if not isinstance(other, VariableName):
-            return False
-        else:
-            return str(self) == str(other)
-
-    def __hash__(
-        self,
-    ):
-        return hash(str(self))
-
-    def __str__(
-        self,
-    ) -> str:
-        return self.__str
-
-    def __repr__(
-        self,
-    ) -> str:
-        return f"{type(self).__name__}({str(self)!r})"
-
-    __to_omc_literal__ = __str__
-
-
-class TypeName(
-):
-    __slots__ = (
-        "__parts",
-    )
-
-    __parts: typing.Tuple[str, ...]
-
-    def __new__(cls, *parts):
-        if len(parts) == 1:
-            part, = parts
-            if isinstance(part, TypeName):
-                return part
-
-        return _TypeName_from_valid_parts_no_check(
-            cls,
-            cls.__checked_parts(
-                parts
-            )
-        )
-
-    @classmethod
-    def from_str(
-        cls,
-        s: str
-    ) -> "TypeName":
-        return parser._TypeName_from_str(s)
-
-    @property
-    def parts(
-        self,
-    ) -> typing.Tuple[str, ...]:
-        return self.__parts
-
-    @property
-    def is_absolute(self) -> bool: return str(self.parts[0]) == "."
-
-    @property
-    def last_identifier(
-        self,
-    ) -> VariableName:
-        return VariableName(self.parts[-1])
-
-    @property
-    def parents(
-        self,
-    ) -> typing.Iterator["TypeName"]:
-        for end in reversed(range(1, len(self.parts))):
-            yield _TypeName_from_valid_parts_no_check(
-                type(self),
-                self.parts[:end],
-            )
-
-    @property
-    def parent(
-        self,
-    ) -> "TypeName":
-        for parent in self.parents:
-            return parent
-        else:
-            return self
-
-    @classmethod
-    def __checked_parts(
-        cls,
-        objs: typing.Iterable,
-    ) -> typing.Iterator[str]:
-
-        def not_checked_parts(
-        ) -> typing.Iterator[str]:
-            for obj in objs:
-                if isinstance(obj, TypeName):
-                    yield from obj.parts
-                elif isinstance(obj, VariableName):
-                    yield str(obj)
-                else:
-                    yield from cls.from_str(str(obj)).parts
-
-        for i, part in enumerate(
-            not_checked_parts()
-        ):
-            if part == "." and not i == 0:
-                raise ValueError(
-                    f"parts[{i}] is invalid, got {part!r}"
-                )
-            yield part
-
-    def __hash__(self):
-        return hash(self.parts)
-
-    def __eq__(self, other):
-        return self.parts == type(self)(other).parts
-
-    def __repr__(
-        self,
-    ) -> str:
-        return f"{type(self).__name__}({str(self)!r})"
-
-    def __str__(
-        self,
-    ) -> str:
-        if self.is_absolute:
-            return self.parts[0] + ".".join(self.parts[1:])
-        else:
-            return ".".join(self.parts)
-
-    __to_omc_literal__ = __str__
-
-    def __truediv__(
-        self,
-        other: typing.Union[str, VariableName, "TypeName"]
-    ):
-        return type(self)(self, other)
-
-
-def _VariableName_from_valid_identifier_no_check(
-    cls: typing.Type["VariableName"],
-    identifier: str,
-):
-    variableName = object.__new__(VariableName)
-    variableName._VariableName__str = identifier
-    return variableName
-
-
-def _TypeName_from_valid_parts_no_check(
-    cls: typing.Type["TypeName"],
-    parts: typing.Iterable[str],
-):
-    typeName = object.__new__(TypeName)
-    typeName._TypeName__parts = tuple(parts)
-    return typeName
-
-
-class Component(
-    typing.NamedTuple,
-):
-    className: TypeName
-    name: VariableName
-    comment: str
-    protected: str
-    isFinal: bool
-    isFlow: bool
-    isStream: bool
-    isReplaceable: bool
-    variability: str
-    innerOuter: str
-    inputOutput: str
-    dimensions: typing.Tuple[str, ...]
-
-
-from . import parser  # noqa: E402
